allow ejecting from coverage reporting in jest_config_template.mjs
This addresses the ability to let users define their own coverage reporting logic in unique cases. For us, we use monocart-coverage-reporter. In doing so, we needed to write our own `globalTeardown` for jest that uses the split coverage logic, but also manipulates the dat files before they are passed to bazel