feat: add `auto_configure_reporting` attribute to allow end users to eject from default behavior (#311)
Close https://github.com/aspect-build/rules_jest/pull/294, use a rule attribute
instead of an ENV variable

This allows us to upstream the pattern we use internally in a way others
can use. We have ejected from the main reporting mechanism and built in
our own because of our needs.